requirements.txt
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.64.1
